Splitting up the database connection creation
Currently there are two ways how to create a database (connection). One: create them directly by using the credeintaials as seen in the constructor of the class. Or two you can also use the mysql connection to creat it. Those are two completely different processes and should have seperate methods.